Allow eBPF to be attached to LSM hooks.
The LSM creates files in securityfs for each hook registered with the
LSM.

    /sys/kernel/security/krsi/<h_name>

The initialization of the hooks is done collectively in an internal
header include/hooks.h

The initialization of the hook for the KRSI LSM:

* Creates the file for the hook in the securityfs.
* Allocates a krsi_hook data structure and stores a pointer to the
  dentry of the newly created file in securityfs.
* A krsi_fsdata is then allocated with a bpf_prog_array which is made
  a part of the d_fsdata of the dentry.

Macros for initialization and cleanup:

- A macro to iterate over each hook (krsi_for_each_hook)
- Provides functionality to cleanup a hook (krsi_free_hook)
- Uses them to handle cleanup during error cases in the initialization
  logic.

A user space program can attach an eBPF program to a security hook by
getting a file descriptor to the respective hook file in securityfs
and using it as a target file descriptor.

When such an attach call is received, the attachment logic looks up the
dentry and appends the program to the bpf_prog_array.

The BPF prorgams are stored in a bpf_prog_array in the d_fsdata,

Change-Id: I0412ed06c7cc680ae80fcd403dfd0573a41257a7